pyproject.toml
[project]
name = "puffin-bench"
version = "0.1.0"
description = "A performance testbench for the puffin fuzzer"
readme = "README.md"
authors = [{ name = "Michael Mera", email = "dev@michaelmera.com" }]
requires-python = ">=3.10"
dependencies = [
    "gitpython>=3.1.43",
    "ipykernel>=6.29.5",
    "jupyter>=1.1.1",
    "jupyterlab>=4.3.1",
    "numpy>=2.1.3",
    "pandas>=2.2.3",
    "plotly>=5.24.1",
    "prefect>=3.1.4",
    "quarto-cli>=1.5.57",
    "quarto>=0.1.0",
    "watchdog>=6.0.0",
    "tabulate>=0.9.0",
    "filelock>=3.17.0",
    "psutil>=6.1.1",
]

[project.scripts]
benchmark = "puffin_bench:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "assertpy>=1.1",
    "pandas-stubs>=2.2.3.241126",
    "pytest>=8.3.4",
    "ruff>=0.7.4",
    "types-psutil>=6.1.0.20241221",
]

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = [
    "RUF",  # Ruff-specific-rules
    "F",    # pyflakes
    "I",    # isort
    "PL",   # pylint
    "UP",   # pyupgrade
    "PGH",  # pygrep-hooks
    "NPY",  # NumPy-specific rules
    "PERF", # Perflint
    "FLY",  # flynt
    "E",    # pycodestyle
    "W",    # pycodestyle
    "YTT",  # flake8-2020
    "B",    # flake8-bugbear
    "Q",    # flake8-quotes
    "T10",  # flake8-debugger
    "INT",  # flake8-gettext
    "LOG",  # flake8-logging
    "PT",   # flake8-pytest-style
    "PIE",  # flake8-pie
    "PYI",  # flake8-pyi
    "TID",  # flake8-tidy-imports
    "ISC",  # flake8-implicit-str-concat
    "TCH",  # flake8-type-checking
    "C4",   # flake8-comprehensions
    "G",    # flake8-logging-format
    "FA",   # flake8-future-annotations
    "SLOT", # flake8-slots
    "RSE",  # flake8-raise"

    # Per-rule selection:

    "S101",   # flake8-bandit: assert (should add tests to exclusion-list)
    "S102",   # flake8-bandit: exec-builtin
    "ICN001", # flake8-import-conventions: unconventional-import-alias
    "TRY002", # tryceratops: raise-vanilla-class
]

ignore = [
    "TCH003", # flake8-type-checking: typing-only-standard-library-import

    # rules confliting with Ruff (as per https://docs.astral.sh/ruff/formatter/#conflicting-lint-rules)
    "W191",
    "E111",
    "E114",
    "E117",
    "D206",
    "D300",
    "Q000",
    "Q001",
    "Q002",
    "Q003",
    "COM812",
    "COM819",
    "ISC001",
    "ISC002",
]
